screens for navigation added

+++ app/src/main/java/com/shift/shiftfinal/data/model/LoanCondition.kt
@@ -2,6 +2,6 @@ package com.shift.shiftfinal.data.model
 
 data class LoanCondition(
     val maxAmount: Int,
-    val percent: Int,
+    val percent: Double,
     val period: Int
 )
