+++ app/src/main/java/com/shift/shiftfinal/domain/entity/LoanConditionEntity.kt
@@ -2,6 +2,6 @@ package com.shift.shiftfinal.domain.entity
 
 data class LoanConditionEntity(
     val maxAmount: Int,
-    val percent: Int,
+    val percent: Double,
     val period: Int
 )
